feat(cluster-tax): close u64 cluster-wealth seam + real-curve M2 sim harness (#626 PR 3/3)
PR 3 of 3 for #626. Two parts:

A. Close the u64 seam (consensus-breaking; rides the protocol-4.0.0 reset).
   Widen the fee/factor/centroid APIs to u128 wealth end-to-end so the
   log-domain curve (PR1 #627) and the u128 accumulator (PR2 #628) are no
   longer throttled by a u64 ceiling:
   - FeeConfig::{cluster_factor, estimate_typical_fee, estimate_fee_*,
     compute_fee*} take u128 cluster_wealth.
   - ring_centroid_implied_factor(&[(u64, u128)]) — the C7 consensus
     fee-floor path (#602) in ledger/store.rs — takes u128 member wealth.
   - Removed PR2's `.min(u64::MAX as u128)` clamps at the three named sites:
     mempool relay fee (mempool.rs), C7 ring-centroid floor (store.rs), and
     the CLI estimate / RPC estimateFee paths (send.rs, rpc/mod.rs).
   - Re-derived the multiply bound each clamp was protecting with explicit
     comments: value(u64) × wealth(u128) can exceed u128 at astronomically
     distant cumulative wealth, so the centroid math uses saturating_mul /
     saturating_add, pinning overflow to u128::MAX → factor 6000 (max floor),
     the conservative anti-gaming direction and identical on every node (no
     fork). Deterministic, no floats.

B. Sim-harness gap closure (spec §7). New library harness
   cluster-tax/src/simulation/m2.rs exercises the REAL production
   ClusterFactorCurve: population declared in BTH, converted via PICO_PER_BTH
   at the curve boundary, per-agent factors derived from simulated cumulative
   cluster wealth through the same compiled factor() the node runs. The legacy
   hardcoded-factor entry point (add_owner_with_factor, the #314 harness) is
   kept but marked deprecated; new work uses add_owner_via_production_curve.
   Implements the #605 run matrix as CLI subcommands:
   - m2-cumulative (run set 1): 10/20yr horizons, honest + gamed, emits
     Δgini vs >0.05, merchant mispricing, whale progressivity.
   - m2-decay (run set 2): epoch-halving decay (half-lives 2/5/10yr, integer
     w >>= 1 per epoch — pure function of height, M3 lesson), plus wash-trading
     evasion (<20% gate) and privacy ring-ID rate (<50% gate) per ANALYSIS.md.
   Documented in --help and experiments/M2_RUNBOOK.md. Full suite is NOT run
   here (#605 owns the empirical runs); tiny-horizon --smoke variants prove
   each run type executes and emits its metrics as unit tests.

Resumes a partial build (previous builder hit its session limit). ZkFeeCurve
re-anchoring already landed in PR1 (#627). Gates: cargo test -p
bth-cluster-tax, cargo test -p botho --lib, cargo build --workspace,
cargo fmt --all -- --check — all green.

+++ cluster-tax/src/fee_curve.rs
@@ -281,7 +281,7 @@ impl FeeConfig {
         &self,
         tx_type: TransactionType,
         tx_size_bytes: usize,
-        cluster_wealth: u64,
+        cluster_wealth: u128,
         num_outputs: usize,
         num_memos: usize,
     ) -> u64 {
@@ -290,7 +290,7 @@ impl FeeConfig {
         }
 
         // Get cluster factor (1x to 6x in 1000-scale fixed point)
-        let cluster_factor = self.cluster_curve.factor(cluster_wealth as u128);
+        let cluster_factor = self.cluster_curve.factor(cluster_wealth);
 
         // Get output penalty (capped quadratic by default)
         let output_penalty = self.output_penalty(num_outputs);
@@ -330,7 +330,7 @@ impl FeeConfig {
         &self,
         tx_type: TransactionType,
         tx_size_bytes: usize,
-        cluster_wealth: u64,
+        cluster_wealth: u128,
         num_memos: usize,
     ) -> u64 {
         // Default to 2 outputs (standard transfer: payment + change)
@@ -342,7 +342,7 @@ impl FeeConfig {
         &self,
         tx_type: TransactionType,
         tx_size_bytes: usize,
-        cluster_wealth: u64,
+        cluster_wealth: u128,
     ) -> u64 {
         self.compute_fee(tx_type, tx_size_bytes, cluster_wealth, 0)
     }
@@ -362,8 +362,8 @@ impl FeeConfig {
     /// Get the cluster factor for a given wealth level.
     ///
     /// Returns the multiplier as a fixed-point value (1000 = 1x, 6000 = 6x).
-    pub fn cluster_factor(&self, cluster_wealth: u64) -> u64 {
-        self.cluster_curve.factor(cluster_wealth as u128)
+    pub fn cluster_factor(&self, cluster_wealth: u128) -> u64 {
+        self.cluster_curve.factor(cluster_wealth)
     }
 
     /// Estimate fee for a typical transaction.
@@ -377,7 +377,7 @@ impl FeeConfig {
     pub fn estimate_typical_fee(
         &self,
         tx_type: TransactionType,
-        cluster_wealth: u64,
+        cluster_wealth: u128,
         num_memos: usize,
     ) -> u64 {
         self.estimate_fee_with_outputs(tx_type, cluster_wealth, 2, num_memos)
@@ -391,7 +391,7 @@ impl FeeConfig {
     pub fn estimate_fee_with_outputs(
         &self,
         tx_type: TransactionType,
-        cluster_wealth: u64,
+        cluster_wealth: u128,
         num_outputs: usize,
         num_memos: usize,
     ) -> u64 {
@@ -419,7 +419,7 @@ impl FeeConfig {
         &self,
         tx_type: TransactionType,
         tx_size_bytes: usize,
-        cluster_wealth: u64,
+        cluster_wealth: u128,
         num_memos: usize,
     ) -> u64 {
         self.compute_fee(tx_type, tx_size_bytes, cluster_wealth, num_memos)
@@ -430,7 +430,7 @@ impl FeeConfig {
         &self,
         tx_type: TransactionType,
         tx_size_bytes: usize,
-        cluster_wealth: u64,
+        cluster_wealth: u128,
         num_outputs: usize,
         num_memos: usize,
     ) -> u64 {
@@ -466,7 +466,7 @@ impl FeeConfig {
         &self,
         tx_type: TransactionType,
         tx_size_bytes: usize,
-        cluster_wealth: u64,
+        cluster_wealth: u128,
         num_memos: usize,
         dynamic_base: u64,
     ) -> u64 {
@@ -502,7 +502,7 @@ impl FeeConfig {
         &self,
         tx_type: TransactionType,
         tx_size_bytes: usize,
-        cluster_wealth: u64,
+        cluster_wealth: u128,
         num_outputs: usize,
         num_memos: usize,
         dynamic_base: u64,
@@ -512,7 +512,7 @@ impl FeeConfig {
         }
 
         // Get cluster factor (1x to 6x in 1000-scale fixed point)
-        let cluster_factor = self.cluster_curve.factor(cluster_wealth as u128);
+        let cluster_factor = self.cluster_curve.factor(cluster_wealth);
 
         // Get output penalty (capped quadratic by default)
         let output_penalty = self.output_penalty(num_outputs);
@@ -539,7 +539,7 @@ impl FeeConfig {
         &self,
         tx_type: TransactionType,
         tx_size_bytes: usize,
-        cluster_wealth: u64,
+        cluster_wealth: u128,
         num_memos: usize,
         dynamic_base: u64,
     ) -> u64 {
@@ -557,7 +557,7 @@ impl FeeConfig {
         &self,
         tx_type: TransactionType,
         tx_size_bytes: usize,
-        cluster_wealth: u64,
+        cluster_wealth: u128,
         num_outputs: usize,
         num_memos: usize,
         dynamic_base: u64,
@@ -1328,8 +1328,12 @@ mod tests {
         );
 
         // Same transaction with a large (1M BTH) cluster → ~5x factor.
-        let fee_large =
-            config.compute_fee(TransactionType::Hidden, 4_000, (1_000_000 * PICO) as u64, 0);
+        let fee_large = config.compute_fee(
+            TransactionType::Hidden,
+            4_000,
+            (1_000_000 * PICO) as u128,
+            0,
+        );
         assert!(
             fee_large > fee_small * 2,
             "Large cluster should pay more: {fee_large} > {fee_small}"
@@ -1435,13 +1439,17 @@ mod tests {
         // Test that fees increase with cluster wealth (picocredit inputs).
         // 1k BTH → 1.265x, 100k BTH (midpoint) → 3.5x, 1M BTH → 5.093x.
         let fee_small =
-            config.compute_fee(TransactionType::Hidden, tx_size, (1_000 * PICO) as u64, 0);
-        let fee_mid =
-            config.compute_fee(TransactionType::Hidden, tx_size, (100_000 * PICO) as u64, 0);
+            config.compute_fee(TransactionType::Hidden, tx_size, (1_000 * PICO) as u128, 0);
+        let fee_mid = config.compute_fee(
+            TransactionType::Hidden,
+            tx_size,
+            (100_000 * PICO) as u128,
+            0,
+        );
         let fee_large = config.compute_fee(
             TransactionType::Hidden,
             tx_size,
-            (1_000_000 * PICO) as u64,
+            (1_000_000 * PICO) as u128,
             0,
         );
